requirements.txt
httpx==0.27.2
openai==1.93.0
python-dotenv==1.1.1



